Extend the usage of ENV and CLIarg
Create a "CLIarguments.js" exclusively for parsing and managing command line arguments.

Create ".env" and write some variables. This is a project meant to be seen by an professional educator so even confidential credentials are or will be exposed.

Implement the use of a bunch of new environment variables and set its default values.

Change server port ENV and CLIargument identifier for better clarity.